Create 2810. Faulty Keyboard
Approach

The given code defines a finalString function that takes a string s as input. It iterates through each character of the input string. If the character is 'i', it reverses the content of the StringBuilder object. If the character is not 'i', it appends the character to the StringBuilder. Finally, it returns the content of the StringBuilder as a string.

Complexity

Time complexity:

The code iterates through each character of the input string exactly once.
The sb.reverse() operation takes linear time based on the length of the content in the StringBuilder.
The overall time complexity is O(n), where n is the length of the input string.

Space complexity:

The code uses a StringBuilder to build the final string.
The space occupied by the StringBuilder is proportional to the length of the input string.
Additionally, there are constant space usages like the char variable c.
Overall, the space complexity is O(n), where n is the length of the input string.