Create 2525. Categorize Box According to Criteria
Approach

The provided solution tackles the problem of categorizing a box based on its dimensions and mass. The code iterates through each dimension and checks whether the box is "Bulky" or "Heavy" based on the given criteria. It then uses the boolean variables bulky and heavy to track these conditions.

The volume of the box is calculated by multiplying its dimensions. If the mass of the box is greater than or equal to 100, the heavy boolean is set to true. If any of the dimensions is greater than or equal to 10,000, or if the calculated volume is greater than or equal to 1,000,000,000, the bulky boolean is set to true.

Finally, the code checks the values of bulky and heavy to determine the appropriate category for the box ("Heavy," "Bulky," both, or neither), and returns the corresponding string.

This solution efficiently solves the problem by performing a fixed number of comparisons and calculations, making it suitable for the given constraints.

Complexity

Time complexity:

The time complexity of this solution is O(1). This is because the code performs a fixed number of comparisons and calculations regardless of the input values. The conditions for checking "Heavy" and "Bulky" are straightforward and don't involve iterations or recursion.

Space complexity:

The space complexity of this solution is O(1) as well. The code uses a constant amount of memory to store the boolean variables bulky and heavy, as well as the volume variable. These variables don't scale with the input, so the space complexity remains constant.