Create 2785. Sort Vowels in a String
Intuition

The given code takes an input string 's' and sorts the vowels (both uppercase and lowercase) in the string in ascending order. It does this by creating a new array 'vowels' to store the vowels from the input string, sorting this array, and then replacing the original characters in 'arr' with the sorted vowels. The function returns the modified string as the output.

Converting input string to a character array (char[] arr) takes O(n) time and O(n) space, where 'n' is the length of the input string 's'.

Counting the number of vowels in the input string takes O(n) time.

Creating a new character array (char[] vowels) to store the vowels takes O(n) space, where 'n' is the number of vowels.

Sorting the vowel array using Arrays.sort takes O(nlogn) time.

Replacing the original characters with sorted vowels in the input array (char[] arr) takes O(n) time.

Approach

This approach is a reasonable way to solve the problem. The function iterates through the input string twice: once to count the vowels and create the 'vowels' array and then to replace the original characters with the sorted vowels. Sorting is performed using the Arrays.sort method, which efficiently handles the sorting task. The overall time and space complexity of the function is acceptable for the problem's constraints.

Complexity

Time complexity:

The overall time complexity of the sortVowels function is dominated by the sorting step, which is O(nlogn).

Space complexity:

The character array 'arr' takes O(n) space.

The character array 'vowels' takes O(n) space, where 'n' is the number of vowels in the input string.

Hence, the overall space complexity of the sortVowels function is O(n).