Create 1103. Distribute Candies to People
Intuition

The given code is a Java implementation of a method distribute Candies, which aims to distribute a given number of candies equally among n people (represented by an array arr).

Approach
ALGORITHM :

The code initializes an integer array arr of size n, where each element of the array represents the number of candies distributed to each person.

It uses two variables i and j. i is used to iterate over the people, and j represents the initial number of candies to be given to the first person, increasing by 1 for each subsequent person.

The code enters a loop that continues until all candies are distributed (i.e., candies is zero).

Within the loop, it checks if i has reached the last person's index. If so, it resets i to 0, effectively distributing candies in a circular manner among the people.

It then checks whether there are enough candies to distribute (candies <= j). If so, it distributes the remaining candies to the current person (arr[i] += candies) and sets candies to 0 to break out of the loop.

If there are more candies than j, it distributes j candies to the current person (arr[i] += j), decreases candies by j, increments i to move to the next person, and increases j for the next iteration.

Complexity

Time complexity:

The time complexity of the code is O(sqrt(candies)), where candies is the total number of candies to be distributed. The loop runs approximately sqrt(candies) times because, in the worst case, the loop iterates up to the square root of the total number of candies.

Space complexity:

The space complexity of the code is O(n) because it uses an integer array arr of size n to store the candies distributed to each person.