Create 2164. Sort Even and Odd Indices Independently
Intuition

It's worth mentioning that bubble sort is not the most efficient sorting algorithm, especially for larger arrays. Other sorting algorithms like merge sort or quicksort have better average and worst-case time complexities. However, if the input size is relatively small or the array is partially sorted, the bubble sort can still provide an acceptable solution

Approach
This code uses bubble sort algorithm to sort the even-indexed and odd-indexed elements separately.

Complexity

Time complexity:

The time complexity of the bubble sort algorithm is O(n^2) in the worst and average case. Since the code uses two separate loops to sort the even-indexed and odd-indexed elements, the overall time complexity will be approximately O(n^2) as well

Space complexity:

The space complexity of the provided code is O(1) since it does not use any additional data structures that grow with the input size